ocean-debris-space (1)/debris-space/requirements.txt
ultralytics==8.4.21
Pillow>=9.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
opencv-python-headless>=4.8.0